feat(dashboard): enhance dashboard functionality with new hooks and refactor settings
- Added `useCompiledColumns` hook to streamline template compilation for column data in `Dashlet` components.
- Refactored `DashletSettings` to utilize `buildPgrestSettingsConfig` and `buildPgrestContentLabels` for improved configuration management.
- Updated imports in various components to reflect the new structure and enhance code clarity.
- Removed redundant state management and logic related to column detection, improving maintainability.
