+++ apps/app/src/features/dashboard/dashlets/data_list/dashlet.settings.tsx
@@ -16,7 +16,8 @@ import { usePgrestSettingsState } from "../common/use-pgrest-settings-state";
 import { PgrestSettingsSection } from "../common/pgrest-settings-section";
 import { TableListSettingsShell } from "../common/table-list-settings-shell";
 import { CheckboxColumnList } from "../common/settings-sections";
-import { fromPgrestParamItems, humanizeKey } from "../common/pgrest-types";
+import { fromPgrestParamItems } from "../common/pgrest-types";
+import { buildPgrestSettingsConfig, buildPgrestContentLabels } from "../common/pgrest-settings-helpers";
 import { tr } from "@/features/i18n/tr.service";
 
 export function DashletSettings({
@@ -42,32 +43,20 @@ export function DashletSettings({
     apiUrl: config.apiUrl,
   });
 
+  // Card layout config (unique to data_list)
+  const cl = config.cardLayout ?? defaultCardLayout;
+  const [titleColumn, setTitleColumn] = useState(cl.titleColumn);
+  const [subtitleColumn, setSubtitleColumn] = useState(cl.subtitleColumn);
+  const [headerBadgeColumns, setHeaderBadgeColumns] = useState<string[]>(cl.headerBadgeColumns);
+  const [kpiColumns, setKpiColumns] = useState<string[]>(cl.kpiColumns);
+  const [footerColumns, setFooterColumns] = useState<string[]>(cl.footerColumns);
+
   const pg = usePgrestSettingsState({
     pgrestFunctionName: config.pgrestFunctionName ?? "",
     pgrestParams: config.pgrestParams ?? [],
     pgrestHttpMethod: config.pgrestHttpMethod ?? "POST",
-    onColumnsDetected: (keys) =>
-      keys.map((key, i) => ({
-        _id: `col-${Date.now()}-${i}`,
-        key: `{{row.${key}}}`,
-        label: humanizeKey(key),
-        type: "text" as const,
-      })),
-    setColumns: s.setColumns,
-    syncFiltersToColumns: (detectedKeys, labelByKey) => {
-      s.setFilterItems((prev) =>
-        prev.map((fi) => {
-          const firstKey = [...detectedKeys][0] ?? "";
-          const column = detectedKeys.has(fi.column) ? fi.column : firstKey;
-          return { ...fi, column, label: labelByKey.get(column) ?? fi.label };
-        }),
-      );
-    },
-    syncSortToColumns: (detectedKeys) => {
-      s.setSortColumns((prev) => prev.filter((k) => detectedKeys.has(k)));
-    },
+    ...buildPgrestSettingsConfig(s),
     onDetectionComplete: (detected) => {
-      // Auto-populate card layout from detected columns
       const keys = detected.map((c) => c.key);
       setTitleColumn(keys[0] ?? "");
       setSubtitleColumn(keys[1] ?? "");
@@ -76,14 +65,6 @@ export function DashletSettings({
       setFooterColumns(keys.slice(9, 11));
     },
   });
-
-  // Card layout config (unique to data_list)
-  const cl = config.cardLayout ?? defaultCardLayout;
-  const [titleColumn, setTitleColumn] = useState(cl.titleColumn);
-  const [subtitleColumn, setSubtitleColumn] = useState(cl.subtitleColumn);
-  const [headerBadgeColumns, setHeaderBadgeColumns] = useState<string[]>(cl.headerBadgeColumns);
-  const [kpiColumns, setKpiColumns] = useState<string[]>(cl.kpiColumns);
-  const [footerColumns, setFooterColumns] = useState<string[]>(cl.footerColumns);
 
   const toggleList = (
     setter: React.Dispatch<React.SetStateAction<string[]>>,
@@ -130,14 +111,7 @@ export function DashletSettings({
   const pgrestContent = (
     <PgrestSettingsSection
       pgrest={pg}
-      labels={{
-        functionName: tr("dashboard.settings.functionName", dictionary),
-        httpMethod: tr("dashboard.settings.httpMethod", dictionary),
-        parameters: tr("dashboard.settings.parameters", dictionary),
-        key: tr("dashboard.settings.key", dictionary),
-        value: tr("common.value", dictionary),
-        addParameter: tr("dashboard.settings.addParameter", dictionary),
-      }}
+      labels={buildPgrestContentLabels(dictionary)}
     />
   );
